# receiver/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Health Receiver
  version: 0.1.0
  description: HealthBeat-compatible, idempotent Apple Health ingest API.
servers:
  - url: http://192.168.1.50:8787
    description: Home LAN
  - url: http://100.64.0.1:8787
    description: Tailscale placeholder
security:
  - bearerAuth: []
paths:
...
  /api/v1/days/{target_date}:
    get:
      operationId: exportAnalysisDay
      parameters:
        - name: target_date
          in: path
          required: true
          schema: {type: string, format: date}
        - name: timezone_name
          in: query
          required: false
          schema: {type: string, default: Asia/Shanghai}
      responses:
        '200':
          description: Minute-grouped quantities, previous-night sleep, target-day workouts, routes, activity summary and freshness.
          content:
            application/json:
              schema: {type: object}
        '401': {$ref: '#/components/responses/Unauthorized'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
...
    Unauthorized:
      description: Missing or invalid token.
...